python/requirements.txt
# Core high-performance event loop
# uvloop>=0.19.0 (LINUX/MacOS USERS ONLY)

# Asynchronous WebSocket client (fastest for pure WS streaming)
websockets>=12.0

# Ultra-fast JSON parser (replaces Python's built-in json)
orjson>=3.10.0

# async HTTP client (REST requests)
aiohttp>=3.10.0

# robust async retry/reconnect/backoff logic
tenacity>=9.0.0

# for timestamping, metrics, and structured logs
loguru>=0.7.2

# for flatbuffers runtime libraries
flatbuffers>=25.9.23

pyzmq>=25.0
# NOTE TO SELF: use python -m pip install -r requirements.txt
